+++ dart/geobase/lib/src/vector_data/model.dart
@@ -1,2 +1,3 @@
 export 'model/bounded.dart';
+export 'model/feature.dart';
 export 'model/geometry.dart';